Move login to home page and update backend port
