+++ frontend/README.md
@@ -1,3 +1,3 @@
 # Frontend
 
-Este directorio contiene un ejemplo muy sencillo de una aplicaci\u00f3n Next.js. Primero se debe iniciar sesi\u00f3n en `http://localhost:8080/users/login` para obtener un token. Despu\u00e9s de autenticarse es posible ver los datos obtenidos desde el backend en la ruta `/consultas`.
+Este directorio contiene un ejemplo muy sencillo de una aplicaci\u00f3n Next.js. La autenticaci\u00f3n se realiza en la p\u00e1gina principal (`/`) y el formulario env\u00eda sus datos a `http://localhost:5000/users/login`. Despu\u00e9s de autenticarse es posible ver los datos obtenidos desde el backend en la ruta `/consultas`.
